+++ src/refactoring/components/AdminPage.tsx
@@ -1,6 +1,6 @@
 import { useState } from "react";
 import { Coupon, Discount, Product } from "../../types.ts";
-import CouponForm from "./admin/CouponForm.tsx";
+import { AddProductForm, CouponForm } from "./admin";
 import Layout from "./shared/Layout.tsx";
 
 interface Props {
@@ -13,20 +13,10 @@ interface Props {
 
 const initialNewDiscount: Discount = { quantity: 0, rate: 0 };
 
-const initialNewProduct: Omit<Product, "id"> = {
-  name: "",
-  price: 0,
-  stock: 0,
-  discounts: [],
-};
-
 export const AdminPage = ({ products, coupons, onProductUpdate, onProductAdd, onCouponAdd }: Props) => {
   const [openProductIds, setOpenProductIds] = useState<Set<string>>(new Set());
   const [editingProduct, setEditingProduct] = useState<Product | null>(null);
 
-  const [showNewProductForm, setShowNewProductForm] = useState(false);
-
-  const [newProduct, setNewProduct] = useState<Omit<Product, "id">>(initialNewProduct);
   const [newDiscount, setNewDiscount] = useState<Discount>(initialNewDiscount);
 
   // No Entity
@@ -50,9 +40,6 @@ export const AdminPage = ({ products, coupons, onProductUpdate, onProductAdd, on
   // ================================================
 
   // Entity
-  const initializeProduct = () => {
-    setNewProduct(initialNewProduct);
-  };
 
   const initializeDiscount = () => {
     setNewDiscount(initialNewDiscount);
@@ -120,72 +107,13 @@ export const AdminPage = ({ products, coupons, onProductUpdate, onProductAdd, on
     }
   };
 
-  const handleAddNewProduct = () => {
-    const productWithId = { ...newProduct, id: Date.now().toString() };
-    onProductAdd(productWithId);
-    initializeProduct();
-    setShowNewProductForm(false);
-  };
-
   return (
     <Layout.Container>
       <Layout.Title>관리자 페이지</Layout.Title>
       <div className="grid grid-cols-1 md:grid-cols-2 gap-6">
         <div>
           <h2 className="text-2xl font-semibold mb-4">상품 관리</h2>
-          <button
-            onClick={() => setShowNewProductForm(!showNewProductForm)}
-            className="bg-green-500 text-white px-4 py-2 rounded mb-4 hover:bg-green-600"
-          >
-            {showNewProductForm ? "취소" : "새 상품 추가"}
-          </button>
-          {showNewProductForm && (
-            <div className="bg-white p-4 rounded shadow mb-4">
-              <h3 className="text-xl font-semibold mb-2">새 상품 추가</h3>
-              <div className="mb-2">
-                <label htmlFor="productName" className="block text-sm font-medium text-gray-700">
-                  상품명
-                </label>
-                <input
-                  id="productName"
-                  type="text"
-                  value={newProduct.name}
-                  onChange={(e) => setNewProduct({ ...newProduct, name: e.target.value })}
-                  className="w-full p-2 border rounded"
-                />
-              </div>
-              <div className="mb-2">
-                <label htmlFor="productPrice" className="block text-sm font-medium text-gray-700">
-                  가격
-                </label>
-                <input
-                  id="productPrice"
-                  type="number"
-                  value={newProduct.price}
-                  onChange={(e) => setNewProduct({ ...newProduct, price: parseInt(e.target.value) })}
-                  className="w-full p-2 border rounded"
-                />
-              </div>
-              <div className="mb-2">
-                <label htmlFor="productStock" className="block text-sm font-medium text-gray-700">
-                  재고
-                </label>
-                <input
-                  id="productStock"
-                  type="number"
-                  value={newProduct.stock}
-                  onChange={(e) => setNewProduct({ ...newProduct, stock: parseInt(e.target.value) })}
-                  className="w-full p-2 border rounded"
-                />
-              </div>
-              <button
-                onClick={handleAddNewProduct}
-                className="w-full bg-blue-500 text-white p-2 rounded hover:bg-blue-600"
-              >
-                추가
-              </button>
-            </div>
-          )}
+          <AddProductForm onSubmit={onProductAdd} />
           <div className="space-y-2">
             {products.map((product, index) => (
               <div key={product.id} data-testid={`product-${index + 1}`} className="bg-white p-4 rounded shadow">
