+++ src/refactoring/components/admin/AddProductForm.tsx
@@ -1,0 +1,95 @@
+import { useCallback, useState } from "react";
+import { Product } from "../../../types";
+
+const initialNewProduct: Omit<Product, "id"> = {
+  name: "",
+  price: 0,
+  stock: 0,
+  discounts: [],
+};
+
+type AddProductFormProps = {
+  onSubmit: (newProduct: Product) => void;
+};
+
+const AddProductForm = ({ onSubmit }: AddProductFormProps) => {
+  const [showNewProductForm, setShowNewProductForm] = useState(false);
+
+  const [newProduct, setNewProduct] = useState<Omit<Product, "id">>(initialNewProduct);
+
+  const initializeProduct = () => {
+    setNewProduct(initialNewProduct);
+  };
+
+  const handleAddNewProduct = () => {
+    const productWithId = { ...newProduct, id: Date.now().toString() };
+    onSubmit(productWithId);
+    initializeProduct();
+    setShowNewProductForm(false);
+  };
+
+  const handleChange = useCallback((e: React.ChangeEvent<HTMLInputElement | HTMLSelectElement>) => {
+    const { name, value } = e.target;
+    setNewProduct({ ...newProduct, [name]: value });
+  }, []);
+
+  return (
+    <>
+      <button
+        onClick={() => setShowNewProductForm(!showNewProductForm)}
+        className="bg-green-500 text-white px-4 py-2 rounded mb-4 hover:bg-green-600"
+      >
+        {showNewProductForm ? "취소" : "새 상품 추가"}
+      </button>
+      {showNewProductForm && (
+        <form className="bg-white p-4 rounded shadow mb-4" onSubmit={handleAddNewProduct}>
+          <h3 className="text-xl font-semibold mb-2">새 상품 추가</h3>
+          <div className="mb-2">
+            <label htmlFor="productName" className="block text-sm font-medium text-gray-700">
+              상품명
+            </label>
+            <input
+              id="productName"
+              type="text"
+              value={newProduct.name}
+              name="name"
+              onChange={handleChange}
+              className="w-full p-2 border rounded"
+            />
+          </div>
+          <div className="mb-2">
+            <label htmlFor="productPrice" className="block text-sm font-medium text-gray-700">
+              가격
+            </label>
+            <input
+              id="productPrice"
+              type="number"
+              value={newProduct.price}
+              name="price"
+              onChange={handleChange}
+              className="w-full p-2 border rounded"
+            />
+          </div>
+          <div className="mb-2">
+            <label htmlFor="productStock" className="block text-sm font-medium text-gray-700">
+              재고
+            </label>
+            <input
+              id="productStock"
+              type="number"
+              value={newProduct.stock}
+              name="stock"
+              onChange={handleChange}
+              className="w-full p-2 border rounded"
+            />
+          </div>
+          <button type="submit" className="w-full bg-blue-500 text-white p-2 rounded hover:bg-blue-600">
+            추가
+          </button>
+        </form>
+      )}
+    </>
+  );
+};
+
+export default AddProductForm;
