+++ src/refactoring/components/admin/index.ts
@@ -1,0 +1,4 @@
+import AddProductForm from "./AddProductForm";
+import CouponForm from "./CouponForm";
+
+export { AddProductForm, CouponForm };
